# v2/packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Infractory control plane",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/agent/v1/enroll": {
      "post": {
        "operationId": "enrollAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollAgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Enrolled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollAgentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent/v1/heartbeat": {
      "post": {
        "operationId": "heartbeat",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Accepted"
          }
        }
      }
    },
    "/agent/v1/tasks/next": {
      "get": {
        "operationId": "nextAgentTask",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Leased task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTask"
                }
              }
            }
          },
          "204": {
            "description": "No task"
          }
        }
      }
    },
    "/agent/v1/nebula-certificate-requests": {
      "post": {
        "operationId": "requestNebulaCertificate",
        "security": [
          {
            "deviceToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NebulaCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed certificate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NebulaCertificateResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EnrollAgentRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "publicKey",
          "capabilities"
        ],
        "properties": {
          "token": {
            "minLength": 32,
            "type": "string"
          },
          "publicKey": {
            "minLength": 32,
            "maxLength": 8192,
            "type": "string"
          },
          "capabilities": {
            "additionalProperties": false,
            "type": "object",
            "required": [
              "os",
              "architecture",
              "dockerVersion",
              "composeVersion",
              "tunAvailable",
              "ntpSynchronized",
              "clockOffsetSeconds"
            ],
            "properties": {
              "os": {
                "anyOf": [
                  {
                    "const": "ubuntu-22.04",
                    "type": "string"
                  },
                  {
                    "const": "ubuntu-24.04",
                    "type": "string"
                  },
                  {
                    "const": "debian-12",
                    "type": "string"
                  }
                ]
              },
              "architecture": {
                "anyOf": [
                  {
                    "const": "amd64",
                    "type": "string"
                  },
                  {
                    "const": "arm64",
                    "type": "string"
                  }
                ]
              },
              "dockerVersion": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "composeVersion": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "tunAvailable": {
                "type": "boolean"
              },
              "ntpSynchronized": {
                "type": "boolean"
              },
              "clockOffsetSeconds": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        }
      },
      "EnrollAgentResponse": {
        "type": "object",
        "required": [
          "nodeId",
          "deviceToken",
          "expiresAt",
          "heartbeatIntervalSeconds"
        ],
        "properties": {
          "nodeId": {
            "format": "uuid",
            "type": "string"
          },
          "deviceToken": {
            "minLength": 32,
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "heartbeatIntervalSeconds": {
            "minimum": 5,
            "type": "integer"
          }
        }
      },
...
      "NebulaCertificateRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "environmentId",
          "publicKey",
          "requestedGroups"
        ],
        "properties": {
          "environmentId": {
            "format": "uuid",
            "type": "string"
          },
          "publicKey": {
            "minLength": 32,
            "maxLength": 8192,
            "type": "string"
          },
          "requestedGroups": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "minLength": 1,
              "maxLength": 255,
              "type": "string"
            }
          }
        }
      },
      "NebulaCertificateResponse": {
        "type": "object",
        "required": [
          "certificate",
          "caCertificate",
          "overlayAddress",
          "expiresAt",
          "blocklist"
        ],
        "properties": {
          "certificate": {
            "minLength": 32,
            "type": "string"
          },
          "caCertificate": {
            "minLength": 32,
            "type": "string"
          },
          "overlayAddress": {
            "minLength": 1,
            "maxLength": 255,
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "blocklist": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "deviceToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  }
}